// server/intelligence-service/src/db/schema.ts
export const issueComment = pgTable(
	"issue_comment",
	{
		// You can use { mode: "bigint" } if numbers are exceeding js number limitations
		id: bigint({ mode: "number" }).primaryKey().notNull(),
		createdAt: timestamp("created_at", { withTimezone: true, mode: "string" }),
		updatedAt: timestamp("updated_at", { withTimezone: true, mode: "string" }),
		authorAssociation: varchar("author_association", { length: 255 }),
		body: text(),
		htmlUrl: varchar("html_url", { length: 255 }),
		// You can use { mode: "bigint" } if numbers are exceeding js number limitations
		authorId: bigint("author_id", { mode: "number" }),
		// You can use { mode: "bigint" } if numbers are exceeding js number limitations
		issueId: bigint("issue_id", { mode: "number" }),
	},
	(table) => [
		foreignKey({
			columns: [table.issueId],
			foreignColumns: [issue.id],
			name: "fk8wy5rxggrte2ntcq80g7o7210",
		}),
		foreignKey({
			columns: [table.authorId],
			foreignColumns: [user.id],
			name: "fkdy6oeojymud1wna20olqgyt31",
		}),
	],
);

export const milestone = pgTable(
	"milestone",
	{
		// You can use { mode: "bigint" } if numbers are exceeding js number limitations
		id: bigint({ mode: "number" }).primaryKey().notNull(),
		createdAt: timestamp("created_at", { withTimezone: true, mode: "string" }),
		updatedAt: timestamp("updated_at", { withTimezone: true, mode: "string" }),
		closedAt: timestamp("closed_at", { withTimezone: true, mode: "string" }),
		description: text(),
		dueOn: timestamp("due_on", { withTimezone: true, mode: "string" }),
		htmlUrl: varchar("html_url", { length: 255 }),
		number: integer().notNull(),
		state: varchar({ length: 255 }),
		title: varchar({ length: 255 }),
		// You can use { mode: "bigint" } if numbers are exceeding js number limitations
		creatorId: bigint("creator_id", { mode: "number" }),
		// You can use { mode: "bigint" } if numbers are exceeding js number limitations
		repositoryId: bigint("repository_id", { mode: "number" }),
	},
	(table) => [
		foreignKey({
			columns: [table.repositoryId],
			foreignColumns: [repository.id],
			name: "fkbjhs37s6qmqtd330gu9mit6w0",
		}),
		foreignKey({
			columns: [table.creatorId],
			foreignColumns: [user.id],
			name: "fkg6ieho7gomiumy85puy6l13f1",
		}),
	],
);
// ...
export const issue = pgTable(
	"issue",
	{
		issueType: varchar("issue_type", { length: 31 }).notNull(),
		// You can use { mode: "bigint" } if numbers are exceeding js number limitations
		id: bigint({ mode: "number" }).primaryKey().notNull(),
		createdAt: timestamp("created_at", { withTimezone: true, mode: "string" }),
		updatedAt: timestamp("updated_at", { withTimezone: true, mode: "string" }),
		body: text(),
		closedAt: timestamp("closed_at", { withTimezone: true, mode: "string" }),
		commentsCount: integer("comments_count").notNull(),
		htmlUrl: varchar("html_url", { length: 255 }),
		isLocked: boolean("is_locked").notNull(),
		number: integer().notNull(),
		state: varchar({ length: 255 }),
		title: varchar({ length: 255 }),
		additions: integer(),
		changedFiles: integer("changed_files"),
		commits: integer(),
		deletions: integer(),
		isDraft: boolean("is_draft"),
		isMergeable: boolean("is_mergeable"),
		isMerged: boolean("is_merged"),
		maintainerCanModify: boolean("maintainer_can_modify"),
		mergeCommitSha: varchar("merge_commit_sha", { length: 255 }),
		mergeableState: varchar("mergeable_state", { length: 255 }),
		mergedAt: timestamp("merged_at", { withTimezone: true, mode: "string" }),
		// You can use { mode: "bigint" } if numbers are exceeding js number limitations
		authorId: bigint("author_id", { mode: "number" }),
		// You can use { mode: "bigint" } if numbers are exceeding js number limitations
		milestoneId: bigint("milestone_id", { mode: "number" }),
		// You can use { mode: "bigint" } if numbers are exceeding js number limitations
		repositoryId: bigint("repository_id", { mode: "number" }),
		// You can use { mode: "bigint" } if numbers are exceeding js number limitations
		mergedById: bigint("merged_by_id", { mode: "number" }),
		hasPullRequest: boolean("has_pull_request").notNull(),
		lastSyncAt: timestamp("last_sync_at", { precision: 6, mode: "string" }),
		badPracticeSummary: text("bad_practice_summary"),
		lastDetectionTime: timestamp("last_detection_time", {
			precision: 6,
			withTimezone: true,
			mode: "string",
		}),
	},
	(table) => [
		foreignKey({
			columns: [table.repositoryId],
			foreignColumns: [repository.id],
			name: "fk76s4b6ncspm9bk35y49xh4s9t",
		}),
		foreignKey({
			columns: [table.milestoneId],
			foreignColumns: [milestone.id],
			name: "fk7t1o4tuel06m9bn4dppqmiod6",
		}),
		foreignKey({
			columns: [table.mergedById],
			foreignColumns: [user.id],
			name: "fkqvnu6vslj5txt8xencru8m6x4",
		}),
		foreignKey({
			columns: [table.authorId],
			foreignColumns: [user.id],
			name: "fkrwr6v8fiqetuiuvfjcvie8s85",
		}),
	],
);

export const label = pgTable(
	"label",
	{
		// You can use { mode: "bigint" } if numbers are exceeding js number limitations
		id: bigint({ mode: "number" }).primaryKey().notNull(),
		color: varchar({ length: 255 }),
		description: varchar({ length: 255 }),
		name: varchar({ length: 255 }),
		// You can use { mode: "bigint" } if numbers are exceeding js number limitations
		repositoryId: bigint("repository_id", { mode: "number" }),
	},
	(table) => [
		foreignKey({
			columns: [table.repositoryId],
			foreignColumns: [repository.id],
			name: "fk2951edbl9g9y8ee1q97e2ff75",
		}),
	],
);
// ...
export const workspace = pgTable("workspace", {
	// You can use { mode: "bigint" } if numbers are exceeding js number limitations
	id: bigint({ mode: "number" }).primaryKey().generatedByDefaultAsIdentity({
		name: "workspace_id_seq",
		startWith: 1,
		increment: 1,
		cache: 1,
	}),
	usersSyncedAt: timestamp("users_synced_at", { precision: 6, mode: "string" }),
});
// ...
export const badPracticeDetection = pgTable(
	"bad_practice_detection",
	{
		// You can use { mode: "bigint" } if numbers are exceeding js number limitations
		id: bigint({ mode: "number" }).primaryKey().generatedByDefaultAsIdentity({
			name: "bad_practice_detection_id_seq",
			startWith: 1,
			increment: 1,
			cache: 1,
		}),
		detectionTime: timestamp("detection_time", {
			precision: 6,
			withTimezone: true,
			mode: "string",
		}),
		summary: text(),
		traceId: varchar("trace_id", { length: 255 }),
		// You can use { mode: "bigint" } if numbers are exceeding js number limitations
		pullrequestId: bigint("pullrequest_id", { mode: "number" }),
	},
	(table) => [
		foreignKey({
			columns: [table.pullrequestId],
			foreignColumns: [issue.id],
			name: "FKhk2vrsr2rdq2gb3cjnvieh3nw",
		}),
	],
);

export const badPracticeFeedback = pgTable(
	"bad_practice_feedback",
	{
		// You can use { mode: "bigint" } if numbers are exceeding js number limitations
		id: bigint({ mode: "number" }).primaryKey().generatedByDefaultAsIdentity({
			name: "bad_practice_feedback_id_seq",
			startWith: 1,
			increment: 1,
			cache: 1,
		}),
		explanation: text(),
		type: varchar({ length: 255 }),
		// You can use { mode: "bigint" } if numbers are exceeding js number limitations
		pullRequestBadPracticeId: bigint("pull_request_bad_practice_id", {
			mode: "number",
		}),
		creationTime: timestamp("creation_time", {
			precision: 6,
			withTimezone: true,
			mode: "string",
		}),
	},
	(table) => [
		foreignKey({
			columns: [table.pullRequestBadPracticeId],
			foreignColumns: [pullrequestbadpractice.id],
			name: "FK34k5tg4qb6gy4g7tn9q8uhogl",
		}),
	],
);
// ...
export const chatMessage = pgTable(
	"chat_message",
	{
		id: uuid().primaryKey().notNull(),
		createdAt: timestamp("created_at", {
			precision: 6,
			withTimezone: true,
			mode: "string",
		}).notNull(),
		metadata: jsonb(),
		role: varchar({ length: 16 }).notNull(),
		parentMessageId: uuid("parent_message_id"),
		threadId: uuid("thread_id").notNull(),
	},
	(table) => [
		foreignKey({
			columns: [table.threadId],
			foreignColumns: [chatThread.id],
			name: "FK8s34d909gxc4xrlvml8gag9kh",
		}),
		foreignKey({
			columns: [table.parentMessageId],
			foreignColumns: [table.id],
			name: "FKd0fewjs0l68rq2bww9h8o4cmb",
		}),
	],
);

export const chatThread = pgTable(
	"chat_thread",
	{
		id: uuid().primaryKey().notNull(),
		createdAt: timestamp("created_at", {
			precision: 6,
			withTimezone: true,
			mode: "string",
		}).notNull(),
		title: text(),
		selectedLeafMessageId: uuid("selected_leaf_message_id"),
		// You can use { mode: "bigint" } if numbers are exceeding js number limitations
		userId: bigint("user_id", { mode: "number" }),
	},
	(table) => [
		foreignKey({
			columns: [table.selectedLeafMessageId],
			foreignColumns: [chatMessage.id],
			name: "FK34beodgwi0g7kn66svlk4hlfr",
		}),
		foreignKey({
			columns: [table.userId],
			foreignColumns: [user.id],
			name: "FKikdxlx9viomcwrgxj7fbyfsew",
		}),
		unique("uc_chat_threadselected_leaf_message_id_col").on(
			table.selectedLeafMessageId,
		),
	],
);

export const issueLabel = pgTable(
	"issue_label",
	{
		// You can use { mode: "bigint" } if numbers are exceeding js number limitations
		issueId: bigint("issue_id", { mode: "number" }).notNull(),
		// You can use { mode: "bigint" } if numbers are exceeding js number limitations
		labelId: bigint("label_id", { mode: "number" }).notNull(),
	},
	(table) => [
		foreignKey({
			columns: [table.issueId],
			foreignColumns: [issue.id],
			name: "fkit5n9c0frugu5m8xqsxtps63r",
		}),
		foreignKey({
			columns: [table.labelId],
			foreignColumns: [label.id],
			name: "fkxbk5rr30kkb6k4ech7x4vh9h",
		}),
		primaryKey({
			columns: [table.issueId, table.labelId],
			name: "issue_label_pkey",
		}),
	],
);

export const issueAssignee = pgTable(
	"issue_assignee",
	{
		// You can use { mode: "bigint" } if numbers are exceeding js number limitations
		issueId: bigint("issue_id", { mode: "number" }).notNull(),
		// You can use { mode: "bigint" } if numbers are exceeding js number limitations
		userId: bigint("user_id", { mode: "number" }).notNull(),
	},
	(table) => [
		foreignKey({
			columns: [table.userId],
			foreignColumns: [user.id],
			name: "fk2cfu8w8wjb9vosy4hbrme0rqe",
		}),
		foreignKey({
			columns: [table.issueId],
			foreignColumns: [issue.id],
			name: "fkocgmsva4p5e8ic9k5dbjqa15u",
		}),
		primaryKey({
			columns: [table.issueId, table.userId],
			name: "issue_assignee_pkey",
		}),
	],
);
// ...
export const chatMessagePart = pgTable(
	"chat_message_part",
	{
		messageId: uuid("message_id").notNull(),
		orderIndex: integer("order_index").notNull(),
		content: jsonb(),
		originalType: varchar("original_type", { length: 128 }),
		type: varchar({ length: 32 }).notNull(),
	},
	(table) => [
		foreignKey({
			columns: [table.messageId],
			foreignColumns: [chatMessage.id],
			name: "FKkfle3niou3f9r63mc3u8vi1na",
		}),
		primaryKey({
			columns: [table.messageId, table.orderIndex],
			name: "chat_message_partPK",
		}),
	],
);

export const document = pgTable(
	"document",
	{
		id: uuid().notNull(),
		versionNumber: integer("version_number").notNull(),
		content: text(),
		createdAt: timestamp("created_at", {
			precision: 6,
			withTimezone: true,
			mode: "string",
		}).notNull(),
		kind: varchar({ length: 255 }).notNull(),
		title: varchar({ length: 255 }).notNull(),
		// You can use { mode: "bigint" } if numbers are exceeding js number limitations
		userId: bigint("user_id", { mode: "number" }).notNull(),
	},
	(table) => [
		index("idx_document_created_at").using(
			"btree",
			table.createdAt.asc().nullsLast().op("timestamptz_ops"),
		),
		index("idx_document_id").using(
			"btree",
			table.id.asc().nullsLast().op("uuid_ops"),
		),
		index("idx_document_user_id").using(
			"btree",
			table.userId.asc().nullsLast().op("int8_ops"),
		),
		foreignKey({
			columns: [table.userId],
			foreignColumns: [user.id],
			name: "FKjhdxdv9sijhujiynqbb5jc010",
		}),
		primaryKey({
			columns: [table.id, table.versionNumber],
			name: "documentPK",
		}),
	],
);